font/convert: convert a font from gimp .c to a more compact format.
Takes a white-on-black 16 character by 16 character image, with one
column of padding to the right and one row of padding below each
character, stored in gimp .c export format.

Converts the image to a format with a lookup table per variable-width
character, and data as row-major bits such that each column is
represented by one or more bytes in data, depending on the character
height.  Columns are byte aligned, and unused rows are set to 0.