libraries: add FHT for calcluating integer DHTs
Note that the Discrete Hartley Transform is not intrinsicly faster
than a good real-to-complex FFT implementation, but the ArduinoFHT
library is indeed considerably faster than the ArduinoFFT one, and
provides the data needed for spectrum display.

Downloaded from http://wiki.openmusiclabs.com/wiki/ArduinoFHT

License legally questionable (non-existent), but openmusiclabs.com
ensures me it's OK to use and distribute with a GPLv3 program.